requirements.txt
flask==2.3.3
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
yfinance==0.2.22
alpaca-py==0.21.0
httpx==0.24.1
beautifulsoup4==4.12.2
lxml==4.9.3
pytz==2023.3
streamlit==1.28.1
edgartools==1.0.1
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0